#15: clarify function names and move testing utils
