+++ tests/test_basic_min_max_via_python.py
@@ -7,7 +7,7 @@ try:
 except ModuleNotFoundError:
     print("module 'mpi4py' is not installed")
 
-from pressiolinalg import utils
+import tests.test_utils as utils
 from pressiolinalg.linalg import _basic_max_via_python
 from pressiolinalg.linalg import _basic_min_via_python
 
@@ -18,7 +18,7 @@ from pressiolinalg.linalg import _basic_min_via_python
 
 def _min_max_setup(operation, rank, out=None, comm=None):
     num_processors = comm.Get_size()
-    local_arr, global_arr = utils.get_local_and_global_arrays(rank, comm)
+    local_arr, global_arr = utils.generate_local_and_global_arrays(rank, comm)
 
     if operation == "min":
         min_result = _basic_min_via_python(local_arr, out=out, comm=comm)
